Add / Error Code and Exception Handling for Authentication

+++ music-api/src/main/java/com/ngohoainam/music_api/service/SongService.java
@@ -9,6 +9,8 @@ import com.ngohoainam.music_api.entity.Album;
 import com.ngohoainam.music_api.entity.Artist;
 import com.ngohoainam.music_api.entity.Song;
 import com.ngohoainam.music_api.entity.User;
+import com.ngohoainam.music_api.exception.AppException;
+import com.ngohoainam.music_api.exception.ErrorCode;
 import com.ngohoainam.music_api.repository.AlbumRepository;
 import com.ngohoainam.music_api.repository.ArtistRepository;
 import com.ngohoainam.music_api.repository.SongRepository;
@@ -28,14 +30,12 @@ public class SongService {
 
     private final SongRepository songRepository;
 
-    private final AlbumRepository albumRepository;
-
     private final ArtistRepository artistRepository;
 
     private final UserRepository userRepository;
     public Song createSong( SongCreateRequest request){
         Artist artist = artistRepository.findById(request.getArtistId()).orElseThrow(()->
-                new RuntimeException("Artist not found"));
+                new AppException(ErrorCode.ARTIST_NOT_FOUND));
 
         Song song = new Song();
         song.setTitle(request.getTitle());
@@ -52,21 +52,26 @@ public class SongService {
         return songRepository.findAll().stream().map(songMapper::toSongResponse).toList();
     }
     public Song getSongById(Long id){
-        return songRepository.findById(id).orElseThrow(()-> new RuntimeException("Song not found"));
+        return songRepository.findById(id).orElseThrow(()-> new AppException(ErrorCode.SONG_NOT_FOUND));
     }
 
     public void deleteSongById(Long id){
-        songRepository.findById(id).orElseThrow(()->new RuntimeException("Song not found"));
+        Song song = songRepository.findById(id).orElseThrow(()->new AppException(ErrorCode.SONG_NOT_FOUND));
+
+        var context = SecurityContextHolder.getContext();
+        String email = context.getAuthentication().getName();
+
+        User currentUser = userRepository.findUserByEmail(email).orElseThrow(()->new AppException(ErrorCode.ARTIST_NOT_FOUND));
+
+        boolean isOwner = song.getArtist().getId().equals(currentUser.getId());
+        boolean isAdmin = currentUser.getRoles().name().equals("ADMIN");
+        if(!isAdmin || !isOwner) throw new AppException(ErrorCode.FORBIDDEN);
         songRepository.deleteById(id);
     }
     public SongResponse updateSongById(Long id, SongUpdateRequest request) {
-        Song song = songRepository.findById(id).orElseThrow(()->new RuntimeException("Song not found"));
-        if (song == null) {
-            return null;
-        }
-        else {
+        Song song = songRepository.findById(id).orElseThrow(()->new AppException(ErrorCode.SONG_NOT_FOUND));
             songMapper.updateSong(song, request);
             return songMapper.toSongResponse(songRepository.save(song));
-        }
+
         }
     }
